+++ google2/CHANGELOG.md
@@ -4,6 +4,8 @@ This file documents changes to the `workbench-google2` library, including notes 
 
 ## 0.36
 
+SBT Dependency: "org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.36-TRAVIS-REPLACE-ME"
+
 ### Changes
 
 - GoogleSubscriberInterpreter (**Breaking**):